config options for ntop
git-svn-id: https://svn.code.sf.net/p/somfree/code/trunk@26 c883dd1d-f74d-40fc-8d12-59f342f64db7

+++ config/src/config.c
@@ -232,8 +232,8 @@ char **argv;
 		while (fgets(buf,sizeof(buf),fp))
 		{
 			const char *p="INVALID_OPTION";
-			int i=strlen(p);
-			int j=strlen(buf)-i;
+			size_t i=strlen(p);
+			int j=(int)(strlen(buf)-i);
 			if (j > 0)
 			{
 				while (j--)
